05-data-platforms/zoomcamp/02-pipeline-mcp/assets/ingestion/requirements.txt
pandas
requests
pyarrow
python-dateutil

